LayoutTests/imported/w3c/web-platform-tests/css/css-images/parsing/object-position-invalid-expected.txt


PASS e.style['object-position'] = "auto" should not set the property value 
PASS e.style['object-position'] = "1px 2px 3px" should not set the property value 
PASS e.style['object-position'] = "left right" should not set the property value 
PASS e.style['object-position'] = "bottom 10%" should not set the property value 
PASS e.style['object-position'] = "bottom 10% top 20%" should not set the property value 
FAIL e.style['object-position'] = "center left 1px" should not set the property value assert_equals: expected "" but got "left 1px center"
FAIL e.style['object-position'] = "center top 2px" should not set the property value assert_equals: expected "" but got "center top 2px"
FAIL e.style['object-position'] = "right 3% center" should not set the property value assert_equals: expected "" but got "right 3% center"
FAIL e.style['object-position'] = "left 4px top" should not set the property value assert_equals: expected "" but got "left 4px top"
FAIL e.style['object-position'] = "right top 5px" should not set the property value assert_equals: expected "" but got "right top 5px"
FAIL e.style['object-position'] = "bottom 6% center" should not set the property value assert_equals: expected "" but got "center bottom 6%"
FAIL e.style['object-position'] = "bottom 7% left" should not set the property value assert_equals: expected "" but got "left bottom 7%"
FAIL e.style['object-position'] = "bottom right 8%" should not set the property value assert_equals: expected "" but got "right 8% bottom"
